requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
loguru==0.7.2
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4